fix(formal): simplify nonlinear accounting checks
- Bound post-epoch reserve quotients above the modeled launch range\n- Assert harvested WETH conservation directly from accounting ledgers\n- Cover temporal reserve-share changes at both boundary allocations

+++ test/formal/StaticsGenesisVault.halmos.t.sol
@@ -29,7 +29,9 @@ contract StaticsGenesisVaultHalmosTest is SymTest, FormalGenesisEnvironment {
         assertEq(duringRedemption.reservePayout, 0);
     }
 
-    function check_epochPurchaseQuoteAfter(uint80 quotient, uint16 remainder) public {
+    /// @dev Quotients through uint64 cover reserves above 100,000 native ETH.
+    ///      Certora separately checks the full-width arithmetic invariant.
+    function check_epochPurchaseQuoteAfter(uint64 quotient, uint16 remainder) public {
         vm.assume(remainder < 5_554);
         uint256 reserve = uint256(quotient) * 5_554 + remainder;
         _donate(reserve);
@@ -42,7 +44,7 @@ contract StaticsGenesisVaultHalmosTest is SymTest, FormalGenesisEnvironment {
         assertEq(afterPurchase.reserveBuyIn, expectedBuyIn);
     }
 
-    function check_epochRedemptionQuoteAfter(uint80 quotient, uint16 remainder) public {
+    function check_epochRedemptionQuoteAfter(uint64 quotient, uint16 remainder) public {
         vm.assume(remainder < 5_555);
         uint256 reserve = uint256(quotient) * 5_555 + remainder;
         _donate(reserve);
